moved relayer claim test from xmrmaker to relayer package
